Fixed duplicate code in controllers' tests
