+++ .idea/.name
@@ -1,1 +1,1 @@
-genequery-kotlin
+genequery
