+++ genequery-rest/src/test/kotlin/gq/rest/api/GeneSetEnrichmentControllerTest.kt
@@ -1,27 +1,21 @@
 package gq.rest.api
 
 import gq.rest.Application
+import gq.rest.api.GeneSetEnrichmentController.Companion.URL
 import org.hamcrest.Matchers.*
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.SpringApplicationConfiguration
-import org.springframework.http.MediaType
-import org.springframework.http.converter.HttpMessageConverter
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter
-import org.springframework.mock.http.MockHttpOutputMessage
 import org.springframework.test.context.TestPropertySource
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner
 import org.springframework.test.context.web.WebAppConfiguration
 import org.springframework.test.web.servlet.MockMvc
-import org.springframework.test.web.servlet.ResultActions
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import org.springframework.test.web.servlet.setup.MockMvcBuilders
 import org.springframework.web.context.WebApplicationContext
-import java.nio.charset.Charset
 
 
 @RunWith(SpringJUnit4ClassRunner::class)
@@ -42,10 +36,6 @@ open class GeneSetEnrichmentControllerTest {
         mockMvc = MockMvcBuilders.webAppContextSetup(wac).build()
     }
 
-    var mappingJackson2HttpMessageConverter: HttpMessageConverter<Any> = MappingJackson2HttpMessageConverter()
-
-    val contentType = MediaType(MediaType.APPLICATION_JSON.type, MediaType.APPLICATION_JSON.subtype, Charset.forName("utf8"));
-
     @Test
     fun testBasicRequestSameSpeciesEntrezToEntrez() {
         val requestForm = GeneSetEnrichmentController.EnrichmentRequestForm()
@@ -58,7 +48,7 @@ open class GeneSetEnrichmentControllerTest {
         requestForm.speciesFrom = "hs"
         requestForm.speciesTo = "hs"
 
-        makeRequest(requestForm)
+        mockMvc.makeRequest(URL, requestForm)
                 .andExpect(status().isOk)
                 .andExpect(jsonPath("$.success", equalTo(true)))
                 .andExpect(jsonPath("$.result.identifiedGeneFormat", equalTo("entrez")))
@@ -96,7 +86,7 @@ open class GeneSetEnrichmentControllerTest {
         requestForm.speciesFrom = "hs"
         requestForm.speciesTo = "hss"
 
-        makeRequest(requestForm)
+        mockMvc.makeRequest(URL, requestForm)
                 .andExpect(status().is4xxClientError)
                 .andExpect(jsonPath("$.success", equalTo(false)))
                 .andExpect(jsonPath("$.result", nullValue()))
@@ -112,7 +102,7 @@ open class GeneSetEnrichmentControllerTest {
         requestForm.speciesFrom = "hs"
         requestForm.speciesTo = "hs"
 
-        makeRequest(requestForm)
+        mockMvc.makeRequest(URL, requestForm)
                 .andExpect(status().is4xxClientError)
                 .andExpect(jsonPath("$.success", equalTo(false)))
                 .andExpect(jsonPath("$.result", nullValue()))
@@ -127,7 +117,7 @@ open class GeneSetEnrichmentControllerTest {
         requestForm.genes = queryGenes
         requestForm.speciesTo = "hs"
 
-        makeRequest(requestForm)
+        mockMvc.makeRequest(URL, requestForm)
                 .andExpect(status().is4xxClientError)
                 .andExpect(jsonPath("$.success", equalTo(false)))
                 .andExpect(jsonPath("$.result", nullValue()))
@@ -141,7 +131,7 @@ open class GeneSetEnrichmentControllerTest {
         requestForm.genes = emptyList()
         requestForm.speciesTo = "hss"
 
-        makeRequest(requestForm)
+        mockMvc.makeRequest(URL, requestForm)
                 .andExpect(status().is4xxClientError)
                 .andExpect(jsonPath("$.success", equalTo(false)))
                 .andExpect(jsonPath("$.result", nullValue()))
@@ -149,18 +139,5 @@ open class GeneSetEnrichmentControllerTest {
                 .andExpect(jsonPath("$.errors", hasItem(containsString("hss"))))
                 .andExpect(jsonPath("$.errors", hasItem(containsString("speciesFrom"))))
                 .andExpect(jsonPath("$.errors", hasItem(containsString("genes"))))
-    }
-
-    private fun makeRequest(form: GeneSetEnrichmentController.EnrichmentRequestForm): ResultActions {
-        return mockMvc.perform(post(GeneSetEnrichmentController.URL).content(json(form)).contentType(contentType))
-                .andDo { handler ->
-                    println(handler.response.contentAsString)
-                }
-    }
-
-    private fun json(o: Any): String {
-        val outputMessage = MockHttpOutputMessage()
-        mappingJackson2HttpMessageConverter.write(o, MediaType.APPLICATION_JSON, outputMessage);
-        return outputMessage.bodyAsString;
     }
 }
